REORG

+++ src/main/java/com/staples/asgard/core/domain/model/price/Autoaddpromotion.java
@@ -1,5 +1,7 @@
 package com.staples.asgard.core.domain.model.price;
 
+import java.math.BigDecimal;
+
 import com.fasterxml.jackson.annotation.JsonAutoDetect;
 import com.fasterxml.jackson.annotation.JsonAutoDetect.Visibility;
 import com.fasterxml.jackson.annotation.JsonInclude;
@@ -7,7 +9,7 @@ import com.fasterxml.jackson.annotation.JsonInclude.Include;
 
 /**
  * @author srani
- *
+ * 
  */
 @JsonAutoDetect(getterVisibility = Visibility.PUBLIC_ONLY, setterVisibility = Visibility.PUBLIC_ONLY, fieldVisibility = Visibility.PUBLIC_ONLY)
 @JsonInclude(Include.NON_NULL)
@@ -22,13 +24,13 @@ public class Autoaddpromotion {
 		return code;
 	}
 
-	private java.lang.Integer couponvalue;
+	private BigDecimal couponvalue;
 
-	public void setCouponValue(java.lang.Integer couponvalue) {
+	public void setCouponValue(BigDecimal couponvalue) {
 		this.couponvalue = couponvalue;
 	}
 
-	public java.lang.Integer getCouponValue() {
+	public BigDecimal getCouponValue() {
 		return couponvalue;
 	}
 }
